Button Displays In Serial Monitor
This one's mine! I organized it a little bit, and I actually know how what all the stuff does now. The actual program has a working stopwatch and a button that when you press it displays: the “button was just pressed!” and “button was just released” into the serial monitor.